edit user hook and its usage
